refactor: Refactoring views.
